fix(ui): move render.js to site/, restore /pkg gitignore, add overlay
The previous commit placed render.js directly in pkg/, but pkg/ is
the wasm-pack/solobase build artifact dir (gitignored). Source files
for pkg/ live in site/ and are copied via solobase.toml's
[[assets.overlay]] entries.

Moved site/render.js, restored /pkg in .gitignore, added the overlay
entry, and added /render.js to extra_bypass_prefix so it's served
without auth. Test import path updated.
