+++ .gitignore
@@ -1,5 +1,6 @@
 /target
 /dist
+/pkg
 /node_modules
 node_modules/
 **/target
